+++ app/src/main/java/com/example/capturethetask/ui/components/CttTopAppBar.kt
@@ -1,0 +1,37 @@
+package com.example.capturethetask.ui.components
+
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.example.capturethetask.R
+
+@Composable
+fun CttTopAppBar(
+    title: String,
+    canNavigateBack: Boolean,
+    modifier: Modifier = Modifier,
+    navigateUp: () -> Unit = {}
+) {
+    if (canNavigateBack) {
+        TopAppBar(
+            title = { Text(title) },
+            modifier = modifier,
+            navigationIcon = {
+                IconButton(onClick = navigateUp) {
+                    Icon(
+                        imageVector = Icons.Filled.ArrowBack,
+                        contentDescription = stringResource(R.string.back_button)
+                    )
+                }
+            }
+        )
+    } else {
+        TopAppBar(title = { Text(title) }, modifier = modifier)
+    }
+}
